[luci] Propagate concat qparam during quantization
This changes PropagateConcatenationQparamPass to a function inside QuantizeWithMinMaxPass

  - To guarantee the validity of graph before/after applying Pass

ONE-DCO-1.0-Signed-off-by: Hyukjin Jeong <hj1.jeong@samsung.com>
