--- compiler/luci/pass/src/PropagateConcatenationQparamPassInternal.h
@@ -1,29 +1,0 @@
-/*
- * Copyright (c) 2020 Samsung Electronics Co., Ltd. All Rights Reserved
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#ifndef __LUCI_CIRCLE_PROPAGATE_CONCATENATION_QPARAM_PASS_INTERNAL_H__
-#define __LUCI_CIRCLE_PROPAGATE_CONCATENATION_QPARAM_PASS_INTERNAL_H__
-
-#include <luci/IR/CircleNodes.h>
-
-namespace luci
-{
-
-void propagate_concat_quantparam(luci::CircleConcatenation *concat);
-
-} // namespace luci
-
-#endif // __LUCI_CIRCLE_PROPAGATE_CONCATENATION_QPARAM_PASS_INTERNAL_H__
